requirements.txt

streamlit
pandas
duckdb
sentence-transformers
faiss-cpu
langgraph
groq
python-dotenv
tabulate